scripted-diff: Avoid `interface` keyword. Rename `interface` dir to `interfaces`
-BEGIN VERIFY SCRIPT-

git mv src/interface src/interfaces
git grep -l "interface/" | xargs sed -i "s,interface/,interfaces/,g";

-END VERIFY SCRIPT-
