+++ src/interfaces/wallet.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
 
-#include "interface/wallet.h"
+#include "interfaces/wallet.h"
 #include "wallet.h"
 
 namespace interfaces {
